Loads data correctly

+++ code/Data.cpp
@@ -30,13 +30,13 @@ void Data::load(const char* filename)
     image.clear();
 
     double temp1, temp2, temp3;
-    size_t temp4;
+    double temp4;
     while(fin>>temp1 && fin>>temp2 && fin>>temp3 && fin>>temp4)
     {
         t.push_back(temp1);
         y.push_back(temp2);
         sig.push_back(temp3);
-        image.push_back(temp4);
+        image.push_back((size_t)temp4);
 
         if(t.size() >= 2 && t.back() < t[t.size() - 2])
             throw std::invalid_argument("Unsorted t-values in file.");
